Added a timeline view
The timeline view allows users to see the range of decades during which
music in the collection was published. The height of the decade in the
timeline indicates how much music in the collection was published
during that period. Selecting a decade shows an index of all the music
published during that decade.

Updated the data to include information about the decade in which the
score was published.

Added an app icon.